cap: split the FP8 weight-install setters to qwen3_ssm/init_fp8.rs
The boot-audit hook pushed init.rs (499 on main) to 506; the setter
trio + predequant move to a sibling with real headroom (424/95).
