+++ crates/spark-model/src/layers/qwen3_ssm/mod.rs
@@ -413,6 +413,7 @@ fn ssm_m128_min_m() -> Option<u32> {
 mod debug;
 pub mod gdn_flags;
 mod init;
+mod init_fp8;
 mod ssm_forward;
 pub(crate) mod ssm_h_fp16;
 mod trait_decode;
